Move post meta logic to Post_Types and add tests
Refactored prevent_empty_post_meta and default_post_metadata methods from Activitypub to Post_Types for better separation of concerns. Updated hooks in Activitypub to use the new Post_Types methods. Added unit tests for these methods in a new Test_Post_Types class, and removed related tests from Test_Activitypub.
